# api/web/docs/openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Bintime TT",
        "description": "HTTP JSON API.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://api.localhost/api",
            "description": "Api Bintime TT"
        }
    ],
    "paths": {
...
        "/users/{id}/tasks": {
            "get": {
                "tags": [
                    "User.Task"
                ],
                "description": "Get the user's task list.",
                "operationId": "UserTaskAllTasks",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "example": "1da2e2c6-e904-4350-9c1e-d8fdcaaf7a99"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "description": "Page Number",
                            "type": "integer"
                        }
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "startedAt",
                                "-startedAt",
                                "title",
                                "-title",
                                "description",
                                "-description"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/App\\OpenApi\\User\\TaskPaginateSchema"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "User.Task"
                ],
                "description": "Create a new task for the user.",
                "operationId": "UserTaskCreateNew",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/App\\Http\\Forms\\User\\User\\Task\\CreateNewForm"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/App\\Http\\Response\\Responder\\User\\User\\TaskResponder"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/App\\Http\\Response\\Responder\\Error\\ViolationResponder"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/App\\Http\\Response\\Responder\\Error\\NotFoundResponder"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/App\\Http\\Response\\Responder\\Error\\EntityResponder"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/users/{id}/tasks/stats": {
            "get": {
                "tags": [
                    "User.Task"
                ],
                "description": "Get stats a task for the user.",
                "operationId": "UserTaskAllStatsByUser",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "example": "1da2e2c6-e904-4350-9c1e-d8fdcaaf7a99"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/App\\Http\\Response\\Responder\\User\\User\\TaskStatsResponder"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/users/{id}/tasks/{taskId}": {
            "put": {
                "tags": [
                    "User.Task"
                ],
                "description": "Update task content for user.",
                "operationId": "UserTaskEdit",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "example": "1da2e2c6-e904-4350-9c1e-d8fdcaaf7a99"
                        }
                    },
                    {
                        "name": "taskId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "example": "1da2e2c6-e904-4350-9c1e-d8fdcaaf7a99"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/App\\Http\\Forms\\User\\User\\Task\\EditForm"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/App\\Http\\Response\\Responder\\User\\User\\TaskResponder"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/App\\Http\\Response\\Responder\\Error\\ViolationResponder"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/App\\Http\\Response\\Responder\\Error\\NotFoundResponder"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/App\\Http\\Response\\Responder\\Error\\EntityResponder"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "User.Task"
                ],
                "description": "Delete a task for a user.",
                "operationId": "UserTaskDelete",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "example": "1da2e2c6-e904-4350-9c1e-d8fdcaaf7a99"
                        }
                    },
                    {
                        "name": "taskId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "example": "1da2e2c6-e904-4350-9c1e-d8fdcaaf7a99"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/App\\Http\\Response\\Responder\\Error\\NotFoundResponder"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/App\\Http\\Response\\Responder\\Error\\EntityResponder"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}